# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Management API
  version: 1.0.0
  summary: The versioned management API — the single door every client (web UI, CLI, MCP server) uses to read and change platform state.
  description: >-
    This document is the source of truth every generated client (TypeScript
    for the web UI, Go for the CLI and MCP server) is produced from. Every
    operation lives under the `/api/v1` prefix. Cross-cutting conventions
    (error shape, optimistic concurrency, pagination, idempotency, trace
    propagation, MCP tag curation, the label-selector grammar, and the
    evolution policy) are specified normatively in `contracts/api-1.md`;
    this document implements those conventions as reusable components and
    applies them to every operation below. `scope-nodes` and `automations`
    are worked fully end to end (schemas, every verb, every error response);
    the remaining operations are SHAPE stubs — path, summary, tag and
    parameters, with the response schema reserved for a later minor.

    Every operation below is served by a real handler, and every route the
    server mounts appears below: the two sets are diffed against each other
    on every build (`internal/app/api`'s surface test), so this document
    cannot promise a generated client an operation that 404s, and a handler
    cannot be added that no client can discover and no schema constrains.
  license:
    name: GNU Affero General Public License v3.0 only
    identifier: AGPL-3.0-only
servers:
...
security:
  - SessionCookie: []
  - ApiKey: []
paths:
...
  /automations:
    get:
      operationId: listAutomations
      summary: List automations
      description: Returns a keyset-paginated, selector-filterable page of automations (rules).
      tags: [automations, "mcp:read"]
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SelectorParam"
        - $ref: "#/components/parameters/TraceIdParam"
      responses:
        "200":
          description: A page of automations.
          headers:
            Trace-Id:
              $ref: "#/components/headers/TraceIdResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      operationId: createAutomation
      summary: Create an automation
      tags: [automations, "mcp:act"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
        - $ref: "#/components/parameters/TraceIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationCreate"
      responses:
        "201":
          description: The created automation.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Trace-Id:
              $ref: "#/components/headers/TraceIdResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Automation"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableContent"
...
  /automations/{automation_id}/run:
    parameters:
      - name: automation_id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Ulid"
    post:
      operationId: runAutomation
      summary: Run an automation now
      description: >-
        Runs the automation's action sequence NOW, for real, subject to its
        declared mode's normal mode-evaluation rules and to its own
        conditions: a `device_command` reaches the physical device down the
        owning relay's connection, and a signage action writes the targeted
        screens' program override. The response reports what it did, target by
        target. Pass `dry_run: true` to evaluate everything and withhold every
        effect. A mutating POST outside plain resource creation — accepts
        Idempotency-Key so a client's retry-on-timeout cannot double-fire the
        run.
      tags: [automations, "mcp:act"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
        - $ref: "#/components/parameters/TraceIdParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationRunRequest"
      responses:
        "200":
          description: The run completed; the response carries its disposition and every effect it produced.
          headers:
            Trace-Id:
              $ref: "#/components/headers/TraceIdResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRunResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableContent"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /automations/bulk-enable:
    post:
      operationId: bulkEnableAutomations
      summary: Enable or disable a selector-matched set of automations
      description: >-
        Fleet-mutating operation
        (`contracts/api-1.md#fleet-mutating-operations--the-job-resource`):
        targets every automation the selector matches, accepts the same
        label-selector grammar a list operation does as its target
        predicate, and returns 202 with a Job resource rather than
        blocking until every target finishes.
      tags: [automations, "mcp:act"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
        - $ref: "#/components/parameters/TraceIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationBulkEnableRequest"
      responses:
        "202":
          description: Accepted; poll the returned Job resource for completion.
          headers:
            Trace-Id:
              $ref: "#/components/headers/TraceIdResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableContent"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  # ---------------------------------------------------------------------
  # devices + entities — the device plane's read surface and its one
  # mutating operation, the command a relay executes against a physical
  # device (relay/1 Device plane, REL-112/113).
  #
  # Both families are read-only here: a device and its entities enter the
  # registry through the relay's own discovery and adoption plane, never
  # through an authoring write on this API, so neither declares a create,
  # update, or delete. Adoption operations are a later minor, additive to
  # these paths.
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Browser session cookie issued by /auth/login. Mutating requests additionally require the platform's CSRF token.
    ApiKey:
      type: http
      scheme: bearer
      bearerFormat: "wv_<random>, scoped to a principal and role"
      description: API-key principal (user-issued or internal-service), presented as a bearer token.
  parameters:
    CursorParam:
      name: cursor
      in: query
      required: false
      description: Opaque continuation token from a prior response's `cursor` field. Never constructed or parsed by the client.
      schema:
        $ref: "#/components/schemas/Cursor"
    LimitParam:
      name: limit
      in: query
      required: false
      description: Maximum rows to return in this page.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    SelectorParam:
      name: selector
      in: query
      required: false
      description: >-
        A label-selector string: comma-separated, ANDed terms (equality,
        inequality, set-membership, set-exclusion, existence,
        non-existence, or a `scope_node subtree <ulid>` term). See
        `contracts/api-1.md#label-selector-grammar` for the full grammar.
      schema:
        type: string
        example: "kind=screen,env in (prod,staging),scope_node subtree 01J8Z2Q1M8H8N4T0V1W2X3Y4Z5"
...
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      required: false
      description: Client-generated opaque replay key, scoped to (principal, method, path). Optional; strongly recommended on any POST a client might retry.
      schema:
        type: string
        minLength: 1
        maxLength: 255
    TraceIdParam:
      name: Trace-Id
      in: header
      required: false
      description: Caller-supplied trace ID (ULID- or UUID-class, 20-36 chars). A non-conforming value is discarded and replaced server-side; the request still proceeds.
      schema:
        $ref: "#/components/schemas/TraceId"
...
  headers:
    ETag:
      description: Strong validator derived solely from the resource's `revision`. Present on every response representing a single mutable resource.
      required: true
      schema:
        type: string
        example: '"3"'
    TraceIdResponse:
      description: The trace ID that correlates this request across every component it touched — either the caller-supplied value (if it validated) or a freshly generated one.
      required: true
      schema:
        $ref: "#/components/schemas/TraceId"

  # Note: the Deprecation (RFC 9745) and Sunset (RFC 8594) response headers
  # this contract's evolution policy uses (contracts/api-1.md, API-082) are
  # not modeled as reusable components here — nothing in this document is
  # deprecated yet, so there is no operation to attach them to without
  # faking a deprecation. contracts/api-1.md is their normative source; a
  # future minor wires them onto whichever operation first deprecates.
  responses:
    BadRequest:
      description: The request was malformed (validation failure, invalid selector, invalid cursor).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Unauthorized:
      description: No valid principal was presented.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: The principal is authenticated but not authorized for this operation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Conflict:
      description: >-
        The request cannot be applied against the resource's current state: an
        Idempotency-Key reused with a different body or still in progress, or a
        deletion refused because something still depends on the row.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    TooManyRequests:
      description: The principal exceeded its rate limit.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Ulid:
      type: string
      description: A 26-character Crockford-base32 identifier, lexicographically sortable by creation time.
      pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
      examples: ["01J8Z3K4N5P6Q7R8S9T0V1W2X3"]
...
    AutomationListResponse:
      type: object
      required: [items, cursor]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Automation"
        cursor:
          $ref: "#/components/schemas/Cursor"
...